[BSM-23] refactor: update authStore to use new mock path
